Move `ESFSignature640` to feature types header
