+++ common/include/pcl/impl/point_types.hpp
@@ -158,7 +158,6 @@ namespace pcl
     namespace traits
     {
       template<> struct descriptorSize<BRISKSignature512> { static constexpr const int value = 64; };
-      template<> struct descriptorSize<ESFSignature640> { static constexpr const int value = 640; };
       template<> struct descriptorSize<GASDSignature512> { static constexpr const int value = 512; };
       template<> struct descriptorSize<GASDSignature984> { static constexpr const int value = 984; };
       template<> struct descriptorSize<GASDSignature7992> { static constexpr const int value = 7992; };
@@ -1099,20 +1098,6 @@ namespace pcl
     friend std::ostream& operator << (std::ostream& os, const BRISKSignature512& p);
   };
 
-  PCL_EXPORTS std::ostream& operator << (std::ostream& os, const ESFSignature640& p);
-  /** \brief A point structure representing the Ensemble of Shape Functions (ESF).
-    * \ingroup common
-    */
-  struct ESFSignature640
-  {
-    float histogram[640] = {0.f};
-    static constexpr int descriptorSize () { return detail::traits::descriptorSize_v<ESFSignature640>; }
-
-    inline constexpr ESFSignature640 () = default;
-
-    friend std::ostream& operator << (std::ostream& os, const ESFSignature640& p);
-  };
-
   PCL_EXPORTS std::ostream& operator << (std::ostream& os, const GASDSignature512& p);
   /** \brief A point structure representing the Globally Aligned Spatial Distribution (GASD) shape descriptor.
   * \ingroup common
@@ -1554,10 +1539,6 @@ POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::BRISKSignature512,
     (float, scale, brisk_scale)
     (float, orientation, brisk_orientation)
     (unsigned char[64], descriptor, brisk_descriptor512)
-)
-
-POINT_CLOUD_REGISTER_POINT_STRUCT (pcl::ESFSignature640,
-    (float[640], histogram, esf)
 )
 
 POINT_CLOUD_REGISTER_POINT_STRUCT(pcl::GASDSignature512,
