+++ common/src/feature_types.cpp
@@ -148,4 +148,12 @@ namespace pcl
     os << (i == 0 ? "(" : "") << p.histogram[i] << (i < 307 ? ", " : ")");
     return (os);
   }
+
+  std::ostream&
+  operator << (std::ostream& os, const ESFSignature640& p)
+  {
+    for (int i = 0; i < 640; ++i)
+    os << (i == 0 ? "(" : "") << p.histogram[i] << (i < 639 ? ", " : ")");
+    return (os);
+  }
 } // namespace pcl
